Data persistence with localStorage, check to enable localStorage, toast errors fixed

+++ src/data/posts.js
@@ -1,9 +1,23 @@
 import axios from "axios"
 
+let useLocalStorage = JSON.parse(localStorage.getItem('useLocalStorage')) ?? true;
+
 export async function getPosts() {
+    // -----  Temporary line of code for data persistence -----
+    const savedPosts = JSON.parse(localStorage.getItem('posts'))
+    if (savedPosts && useLocalStorage) {
+        return savedPosts
+    }
+    // -----  Temporary line of code for data persistence -----
     const results = await axios.get(import.meta.env.VITE_POST_API_URL)
         .then((response) => {
-            return response.data
+            const posts = response.data
+            // -----  Temporary line of code for data persistence -----
+            if (useLocalStorage) {
+                localStorage.setItem('posts', JSON.stringify(posts))
+            }
+            // -----  Temporary line of code for data persistence -----
+            return posts
         }).catch((error) => {
             console.log(error)
             return []
@@ -12,6 +26,13 @@ export async function getPosts() {
 }
 
 export async function getPost(id) {
+    // -----  Temporary line of code for data persistence -----
+    const savedPosts = JSON.parse(localStorage.getItem('posts'))
+    if (savedPosts && savedPosts.length > 0 && useLocalStorage) {
+        const postInArray = savedPosts.filter(post => post.id === Number(id))
+        return (Object.keys(postInArray).length > 0) ? postInArray[0] : null
+    }
+    // -----  Temporary line of code for data persistence -----
     const results = await axios.get(`${import.meta.env.VITE_POST_API_URL}/${id}`)
         .then((response) => {
             return response.data
@@ -24,23 +45,19 @@ export async function getPost(id) {
 
 export async function createPost(post) {
     const { title, body } = post
+    // -----  Temporary line of code for data persistence -----
+    const savedPosts = JSON.parse(localStorage.getItem('posts'))
+    if (savedPosts && Object.keys(post).length > 0 && useLocalStorage) {
+        const newIndex = savedPosts.length + 1
+        post.id = (newIndex) // Becase the API always send 101 and we need unique index
+        savedPosts[newIndex - 1] = post
+        localStorage.setItem('posts', JSON.stringify(savedPosts))
+    }
+    // -----  Temporary line of code for data persistence -----
     const results = await axios.post(import.meta.env.VITE_POST_API_URL, {
         title: title,
         body: body,
         userId: 1
-    }).then((response) => {
-        return response.data;
-    }).catch((error) => {
-        console.log(error)
-        return null
-    });
-    return results
-}
-
-export async function updatePost(id, data) {
-    const results = await axios.put(`${import.meta.env.VITE_POST_API_URL}/${id}`, {
-        title: data.title,
-        body: data.body
     }).then((response) => {
         return response.data
     }).catch((error) => {
@@ -50,7 +67,45 @@ export async function updatePost(id, data) {
     return results
 }
 
+export async function updatePost(id, data) {
+    // -----  Temporary line of code for data persistence -----
+    const savedPosts = JSON.parse(localStorage.getItem('posts'))
+    if (savedPosts && savedPosts.length > 0 && Object.keys(data).length > 0 && useLocalStorage) {
+        const postInArray = savedPosts.filter(post => post.id === Number(id))
+        if (Object.keys(postInArray).length > 0) {
+            const postToUpdate = postInArray[0]
+            const indexToUdate = data.id - 1
+            postToUpdate.id = Number(data.id)
+            postToUpdate.title = data.title
+            postToUpdate.body = data.body
+            savedPosts[indexToUdate] = postToUpdate
+            localStorage.setItem('posts', JSON.stringify(savedPosts))
+        }
+        return data;
+    }
+    // -----  Temporary line of code for data persistence -----
+    const results = await axios.put(`${import.meta.env.VITE_POST_API_URL}/${id}`, {
+        title: data.title,
+        body: data.body
+    }).then((response) => {
+        const post = response.data
+        return post
+    }).catch((error) => {
+        console.log(error)
+        return null
+    });
+    return results
+}
+
 export async function deletePost(id) {
+    // -----  Temporary line of code for data persistence -----
+    const savedPosts = JSON.parse(localStorage.getItem('posts'))
+    if (savedPosts && savedPosts.length > 0 && useLocalStorage) {
+        const postsUpdated = savedPosts.filter(post => post.id !== Number(id))
+        localStorage.setItem('posts', JSON.stringify(postsUpdated))
+        return true;
+    }
+    // -----  Temporary line of code for data persistence -----
     const results = await axios.delete(`${import.meta.env.VITE_POST_API_URL}/${id}`)
         .then(() => {
             return true
